move to os/mod.rs, lint linux, add test
